+++ novu/api/message.py
@@ -5,9 +5,9 @@ from typing import Dict, Optional, Union
 
 import requests
 
-from novu.api.base import Api
+from novu.api.base import Api, PaginationIterator
 from novu.constants import MESSAGES_ENDPOINT
-from novu.dto.message import PaginatedMessageDto
+from novu.dto.message import MessageDto, PaginatedMessageDto
 
 
 class MessageApi(Api):
@@ -25,7 +25,12 @@ class MessageApi(Api):
         self._message_url = f"{self._url}{MESSAGES_ENDPOINT}"
 
     def list(
-        self, limit: int = 10, page: int = 0, channel: Optional[str] = None, subscriber_id: Optional[str] = None
+        self,
+        limit: int = 10,
+        page: int = 0,
+        channel: Optional[str] = None,
+        subscriber_id: Optional[str] = None,
+        transaction_id: Optional[str] = None,
     ) -> PaginatedMessageDto:
         """List messages
 
@@ -34,6 +39,7 @@ class MessageApi(Api):
             page: The page to fetch, defaults to 0
             channel: The channel for the messages you wish to list. Defaults to None.
             subscriber_id: The subscriberId for the subscriber you like to list messages for
+            transaction_id: The transactionId for the messages you wish to list. Defaults to None.
 
         Returns:
             Returned a paginated struct containing retrieved messages
@@ -44,8 +50,31 @@ class MessageApi(Api):
             payload["channel"] = channel
         if subscriber_id:
             payload["subscriberId"] = subscriber_id
+        if transaction_id:
+            payload["transactionId"] = transaction_id
 
         return PaginatedMessageDto.from_camel_case(self.handle_request("GET", self._message_url, payload=payload))
+
+    def stream(
+        self, channel: Optional[str] = None, subscriber_id: Optional[str] = None
+    ) -> PaginationIterator[MessageDto]:
+        """Stream all existing messages into an iterator.
+
+        Args:
+            channel: The channel for the messages you wish to list. Defaults to None.
+            subscriber_id: The subscriberId for the subscriber you like to list messages for
+
+        Returns:
+            An iterator on all messages available.
+        """
+        payload = {}
+
+        if channel:
+            payload["channel"] = channel
+        if subscriber_id:
+            payload["subscriberId"] = subscriber_id
+
+        return PaginationIterator(self, MessageDto, self._message_url, payload=payload)
 
     def delete(self, message_id: str) -> bool:
         """Deletes a message entity from the Novu platform