+++ tests/api/test_event.py
@@ -9,6 +9,8 @@ from tests.factories import MockResponse
 
 
 class EventApiTests(TestCase):
+    api: EventApi
+
     @classmethod
     def setUpClass(cls) -> None:
         NovuConfig.configure("sample.novu.com", "api-key")
@@ -17,14 +19,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_with_single_recipient(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.trigger("test-template", {"subscriber_id": "sample-recipient"}, {})
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -38,14 +40,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_with_multiple_recipients(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.trigger("test-template", ["sample-recipient-1", "sample-recipient-2"], {})
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -63,14 +65,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_with_overrides(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.trigger("test-template", "sample-recipient", {}, {"an": "override"})
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -89,14 +91,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_with_actor(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.trigger("test-template", "sample-recipient", {}, actor="actor-id")
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -115,14 +117,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_with_transaction_id(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.trigger("test-template", "sample-recipient", {}, transaction_id="sample-test")
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -139,9 +141,30 @@ class EventApiTests(TestCase):
         )
 
     @mock.patch("requests.request")
+    def test_trigger_with_tenant(self, mock_request: mock.MagicMock) -> None:
+        mock_request.return_value = MockResponse(
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
+        )
+
+        result = self.api.trigger("test-template", "sample-recipient", {}, tenant="tenant-id")
+
+        self.assertIsInstance(result, EventDto)
+        self.assertTrue(result.acknowledged)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
+        self.assertEqual(result.transaction_id, "sample-test")
+        mock_request.assert_called_once_with(
+            method="POST",
+            url="sample.novu.com/v1/events/trigger",
+            headers={"Authorization": "ApiKey api-key"},
+            json={"name": "test-template", "to": "sample-recipient", "payload": {}, "tenant": "tenant-id"},
+            params=None,
+            timeout=5,
+        )
+
+    @mock.patch("requests.request")
     def test_trigger_topic_with_single_topic(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         topic = TriggerTopicDto("topic-key", "type")
@@ -149,7 +172,7 @@ class EventApiTests(TestCase):
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -167,7 +190,7 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_topic_with_multiple_topics(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         topic_1 = TriggerTopicDto("topic-key-1", "type")
@@ -176,7 +199,7 @@ class EventApiTests(TestCase):
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -194,7 +217,7 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_topic_with_overrides(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         topic = TriggerTopicDto("topic-key", "type")
@@ -202,7 +225,7 @@ class EventApiTests(TestCase):
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -221,7 +244,7 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_topic_with_actor(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         topic = TriggerTopicDto("topic-key", "type")
@@ -229,7 +252,7 @@ class EventApiTests(TestCase):
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -248,7 +271,7 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_trigger_topic_with_transaction_id(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         topic = TriggerTopicDto("topic-key", "type")
@@ -256,7 +279,7 @@ class EventApiTests(TestCase):
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -273,12 +296,39 @@ class EventApiTests(TestCase):
         )
 
     @mock.patch("requests.request")
+    def test_trigger_topic_with_tenant(self, mock_request: mock.MagicMock) -> None:
+        mock_request.return_value = MockResponse(
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
+        )
+
+        topic = TriggerTopicDto("topic-key", "type")
+        result = self.api.trigger_topic("test-template", topic, {}, tenant="tenant-id")
+
+        self.assertIsInstance(result, EventDto)
+        self.assertTrue(result.acknowledged)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
+        self.assertEqual(result.transaction_id, "sample-test")
+        mock_request.assert_called_once_with(
+            method="POST",
+            url="sample.novu.com/v1/events/trigger",
+            headers={"Authorization": "ApiKey api-key"},
+            json={
+                "name": "test-template",
+                "to": [{"topicKey": "topic-key", "type": "type"}],
+                "payload": {},
+                "tenant": "tenant-id",
+            },
+            params=None,
+            timeout=5,
+        )
+
+    @mock.patch("requests.request")
     def test_trigger_bulk_with_one_event(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
             201,
             {
                 "data": [
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"},
                 ]
             },
         )
@@ -294,7 +344,7 @@ class EventApiTests(TestCase):
         triggered_event = result[0]
         self.assertIsInstance(triggered_event, EventDto)
         self.assertTrue(triggered_event.acknowledged)
-        self.assertEqual(triggered_event.status, EventStatus.PROCESSED.value)
+        self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
         self.assertEqual(triggered_event.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -317,8 +367,8 @@ class EventApiTests(TestCase):
             201,
             {
                 "data": [
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[0]},
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[1]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[0]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[1]},
                 ]
             },
         )
@@ -334,7 +384,7 @@ class EventApiTests(TestCase):
         for triggered_event in result:
             self.assertIsInstance(triggered_event, EventDto)
             self.assertTrue(triggered_event.acknowledged)
-            self.assertEqual(triggered_event.status, EventStatus.PROCESSED.value)
+            self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
             self.assertIn(triggered_event.transaction_id, transaction_ids)
 
         mock_request.assert_called_once_with(
@@ -359,8 +409,8 @@ class EventApiTests(TestCase):
             201,
             {
                 "data": [
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[0]},
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[1]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[0]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[1]},
                 ]
             },
         )
@@ -377,7 +427,7 @@ class EventApiTests(TestCase):
         for triggered_event in result:
             self.assertIsInstance(triggered_event, EventDto)
             self.assertTrue(triggered_event.acknowledged)
-            self.assertEqual(triggered_event.status, EventStatus.PROCESSED.value)
+            self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
             self.assertIn(triggered_event.transaction_id, transaction_ids)
 
         mock_request.assert_called_once_with(
@@ -402,8 +452,8 @@ class EventApiTests(TestCase):
             201,
             {
                 "data": [
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[0]},
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[1]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[0]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[1]},
                 ]
             },
         )
@@ -419,7 +469,7 @@ class EventApiTests(TestCase):
         for triggered_event in result:
             self.assertIsInstance(triggered_event, EventDto)
             self.assertTrue(triggered_event.acknowledged)
-            self.assertEqual(triggered_event.status, EventStatus.PROCESSED.value)
+            self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
             self.assertIn(triggered_event.transaction_id, transaction_ids)
 
         mock_request.assert_called_once_with(
@@ -444,8 +494,8 @@ class EventApiTests(TestCase):
             201,
             {
                 "data": [
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[0]},
-                    {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": transaction_ids[1]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[0]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[1]},
                 ]
             },
         )
@@ -464,7 +514,7 @@ class EventApiTests(TestCase):
         for triggered_event in result:
             self.assertIsInstance(triggered_event, EventDto)
             self.assertTrue(triggered_event.acknowledged)
-            self.assertEqual(triggered_event.status, EventStatus.PROCESSED.value)
+            self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
             self.assertIn(triggered_event.transaction_id, transaction_ids)
 
         mock_request.assert_called_once_with(
@@ -487,16 +537,59 @@ class EventApiTests(TestCase):
         )
 
     @mock.patch("requests.request")
+    def test_trigger_bulk_with_multiple_events_and_tenant(self, mock_request: mock.MagicMock) -> None:
+        transaction_ids = ["sample-test", "another-sample-test"]
+
+        mock_request.return_value = MockResponse(
+            201,
+            {
+                "data": [
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[0]},
+                    {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": transaction_ids[1]},
+                ]
+            },
+        )
+
+        events = [
+            InputEventDto(name="test-template", recipients="recipient_1", payload={}, tenant="tenant-id"),
+            InputEventDto(name="test-template", recipients="recipient_2", payload={}),
+        ]
+
+        result = self.api.trigger_bulk(events)
+
+        self.assertEqual(len(result), len(events))
+
+        for triggered_event in result:
+            self.assertIsInstance(triggered_event, EventDto)
+            self.assertTrue(triggered_event.acknowledged)
+            self.assertEqual(triggered_event.status, EventStatus.PROCESSED)
+            self.assertIn(triggered_event.transaction_id, transaction_ids)
+
+        mock_request.assert_called_once_with(
+            method="POST",
+            url="sample.novu.com/v1/events/trigger/bulk",
+            headers={"Authorization": "ApiKey api-key"},
+            json={
+                "events": [
+                    {"name": "test-template", "to": "recipient_1", "payload": {}, "tenant": "tenant-id"},
+                    {"name": "test-template", "to": "recipient_2", "payload": {}},
+                ]
+            },
+            params=None,
+            timeout=5,
+        )
+
+    @mock.patch("requests.request")
     def test_broadcast_with_overrides(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.broadcast("test-template", {}, {"an": "override"})
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -514,14 +607,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_broadcast_with_actor(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.broadcast("test-template", {}, actor="actor-id")
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -539,14 +632,14 @@ class EventApiTests(TestCase):
     @mock.patch("requests.request")
     def test_broadcast_with_transaction_id(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(
-            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED.value, "transactionId": "sample-test"}}
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
         )
 
         result = self.api.broadcast("test-template", {}, transaction_id="sample-test")
 
         self.assertIsInstance(result, EventDto)
         self.assertTrue(result.acknowledged)
-        self.assertEqual(result.status, EventStatus.PROCESSED.value)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
         self.assertEqual(result.transaction_id, "sample-test")
         mock_request.assert_called_once_with(
             method="POST",
@@ -562,10 +655,35 @@ class EventApiTests(TestCase):
         )
 
     @mock.patch("requests.request")
+    def test_broadcast_with_tenant(self, mock_request: mock.MagicMock) -> None:
+        mock_request.return_value = MockResponse(
+            201, {"data": {"acknowledged": True, "status": EventStatus.PROCESSED, "transactionId": "sample-test"}}
+        )
+
+        result = self.api.broadcast("test-template", {}, tenant="tenant-id")
+
+        self.assertIsInstance(result, EventDto)
+        self.assertTrue(result.acknowledged)
+        self.assertEqual(result.status, EventStatus.PROCESSED)
+        self.assertEqual(result.transaction_id, "sample-test")
+        mock_request.assert_called_once_with(
+            method="POST",
+            url="sample.novu.com/v1/events/trigger/broadcast",
+            headers={"Authorization": "ApiKey api-key"},
+            json={
+                "name": "test-template",
+                "payload": {},
+                "tenant": "tenant-id",
+            },
+            params=None,
+            timeout=5,
+        )
+
+    @mock.patch("requests.request")
     def test_delete(self, mock_request: mock.MagicMock) -> None:
         mock_request.return_value = MockResponse(204)
 
-        self.assertIsNone(self.api.delete("sample-test"))
+        self.assertIsNone(self.api.delete("sample-test"))  # type: ignore
 
         mock_request.assert_called_once_with(
             method="DELETE",