+++ CHANGELOG.md
@@ -1,3 +1,101 @@
+# [1.10.0](https://github.com/novuhq/novu-python/compare/v1.9.1...v1.10.0) (2023-11-01)
+
+
+### Bug Fixes
+
+* [#148](https://github.com/novuhq/novu-python/issues/148) make webhook_url and channel optional in SubscriberChannelSettingsCredentialsDto ([a334e48](https://github.com/novuhq/novu-python/commit/a334e4874bfa4ef8b7cfe57f5de696045c568bd3))
+
+
+### Features
+
+* **api:** add transaction id param to  MessageApi list method ([4fc5dad](https://github.com/novuhq/novu-python/commit/4fc5dad80f7d2ecbb6a8fc983f99d30654bae874))
+* **deps:** update dependency sentry-sdk to v1.33.0 ([939e10d](https://github.com/novuhq/novu-python/commit/939e10de2c9c546341cb8c51a197871622e68d4c))
+* **deps:** update dependency sentry-sdk to v1.33.1 ([7d0f327](https://github.com/novuhq/novu-python/commit/7d0f327183fdd06eaac960e662d9ddbcc27f0d48))
+* **deps:** update dependency sphinx to v7.2.6 ([f430a40](https://github.com/novuhq/novu-python/commit/f430a40a68d1d0b73bab1ed97ec445537de67bfa))
+
+## [1.9.1](https://github.com/novuhq/novu-python/compare/v1.9.0...v1.9.1) (2023-10-27)
+
+
+### Bug Fixes
+
+* **subscriber:** prevent errors using bulk_create ([5d29027](https://github.com/novuhq/novu-python/commit/5d290271d24bc2441ccdd9d9adf5ac7fd2f41dda))
+
+# [1.9.0](https://github.com/novuhq/novu-python/compare/v1.8.0...v1.9.0) (2023-10-27)
+
+
+### Features
+
+* added missing delete topic method in TopicApi ([2e1219d](https://github.com/novuhq/novu-python/commit/2e1219d5c408e2638e43cb68111b51daf9341d18))
+* **deps:** update abatilo/actions-poetry action to v2.3.0 ([13dc292](https://github.com/novuhq/novu-python/commit/13dc292ea0131ef0202c1f22e6159d5246623339))
+* **deps:** update actions/checkout action to v4 ([0bc3aff](https://github.com/novuhq/novu-python/commit/0bc3affd1066b4eefbf80ff990c8ae5be60a38c2))
+* **deps:** update actions/setup-python action to v4 ([c76ed15](https://github.com/novuhq/novu-python/commit/c76ed15e3b9fc4322dac51d1824cd505a961df66))
+* **deps:** update dependency black to v23.10.0 ([1cd0805](https://github.com/novuhq/novu-python/commit/1cd0805f68c6353c4004eea0198ab3289196b0a0))
+* **deps:** update dependency black to v23.10.1 ([2651b2c](https://github.com/novuhq/novu-python/commit/2651b2c3415d66e0ef403ee9ba147d39e9601d92))
+* **deps:** update dependency mypy to v1.6.1 ([c7d7899](https://github.com/novuhq/novu-python/commit/c7d7899a5cb20aa551e66662bc1df2979dc1c738))
+* **deps:** update dependency pre-commit to v3 ([a0d7020](https://github.com/novuhq/novu-python/commit/a0d7020903c98f49fa9c686b9407dadf3314dede))
+* **deps:** update dependency pylint to v3 ([2739d9e](https://github.com/novuhq/novu-python/commit/2739d9e41f8f03e196ad72ff827fe50b570cc85d))
+* **deps:** update dependency pylint to v3.0.2 ([e6289ac](https://github.com/novuhq/novu-python/commit/e6289acbade6f508aa56d1110ca117a87360b965))
+* **deps:** update dependency pytest to v7.4.3 ([8fa399f](https://github.com/novuhq/novu-python/commit/8fa399fa5a3260bfc05072f0edb8f712355e1872))
+* **deps:** update dependency sphinx to v7 ([4b55b00](https://github.com/novuhq/novu-python/commit/4b55b00e3f8f1df14d2e1e8fb268a4b65afee78a))
+* **deps:** update dependency sphinxcontrib-mermaid to ^0.9.0 ([0c1aeca](https://github.com/novuhq/novu-python/commit/0c1aeca37716b240cf857b5654e0b29fd4476873))
+* **deps:** update pre-commit hook alessandrojcm/commitlint-pre-commit-hook to v9.5.0 ([fdf285e](https://github.com/novuhq/novu-python/commit/fdf285e428c80994f1d25400833d3efdc71bc80a))
+* **deps:** update pre-commit hook ambv/black to v23 ([6c320c9](https://github.com/novuhq/novu-python/commit/6c320c92f29a688bfa13f9c11deb45cf5602eb08))
+* **deps:** update pre-commit hook ambv/black to v23.10.0 ([78cc46a](https://github.com/novuhq/novu-python/commit/78cc46aa30d5d65aa0cce8230506fc0036880dda))
+* **deps:** update pre-commit hook ambv/black to v23.10.1 ([4c53350](https://github.com/novuhq/novu-python/commit/4c53350bc1866fd0fe15f69a7bd2926f13f41a13))
+* **deps:** update pre-commit hook asottile/pyupgrade to v3.15.0 ([824a0fd](https://github.com/novuhq/novu-python/commit/824a0fdf5ebdef5f7d736960e718e265a394ad64))
+* **deps:** update pre-commit hook asottile/pyupgrade to v3.3.2 ([30df8a9](https://github.com/novuhq/novu-python/commit/30df8a96f2d890fe7d9c6d6f2d5bfece32166b71))
+* **deps:** update pre-commit hook codespell-project/codespell to v2.2.6 ([51b3448](https://github.com/novuhq/novu-python/commit/51b3448d54e183e0d46452c9fb2651db58edf458))
+* **deps:** update pre-commit hook lucas-c/pre-commit-hooks-safety to v1.3.2 ([e122efa](https://github.com/novuhq/novu-python/commit/e122efa6faabadf2cf9ab9079adfe9133b623878))
+* **deps:** update pre-commit hook pre-commit/mirrors-mypy to v1.0.1 ([0e4c5ca](https://github.com/novuhq/novu-python/commit/0e4c5ca47eca79cb4efae37f2882662a843e99b5))
+* **deps:** update pre-commit hook pre-commit/mirrors-mypy to v1.6.0 ([5d1916a](https://github.com/novuhq/novu-python/commit/5d1916a371557701799e09ed292a4dcde0f18010))
+* **deps:** update pre-commit hook pre-commit/mirrors-mypy to v1.6.1 ([029e0c7](https://github.com/novuhq/novu-python/commit/029e0c708fe0c859765a4371a7118b89ee1dab83))
+* **deps:** update pre-commit hook pre-commit/pre-commit-hooks to v4.5.0 ([939836f](https://github.com/novuhq/novu-python/commit/939836f761057833d32aa9126ff78110a00a90dc))
+* **deps:** update pre-commit hook pre-commit/pygrep-hooks to v1.10.0 ([fe1d9cf](https://github.com/novuhq/novu-python/commit/fe1d9cf2cb06522d4d8c8a54f12b70a3b3bc4ecb))
+* **deps:** update pre-commit hook pycqa/bandit to v1.7.5 ([58cd4d4](https://github.com/novuhq/novu-python/commit/58cd4d41b6bce6b3b3f497f465de29f694816ae8))
+
+# [1.8.0](https://github.com/novuhq/novu-python/compare/v1.7.0...v1.8.0) (2023-10-16)
+
+
+### Bug Fixes
+
+* upgrade dependencies for dev environnements ([064e6d0](https://github.com/novuhq/novu-python/commit/064e6d0f13b9777e1f2b07e57472c889b07e9aff))
+
+
+### Features
+
+* added missing regenerate api key method in Environments ([#99](https://github.com/novuhq/novu-python/issues/99)) ([01585c9](https://github.com/novuhq/novu-python/commit/01585c947e9a8cedd9fb40da67c4f44657bdc2c6))
+* **integration:** [#70](https://github.com/novuhq/novu-python/issues/70) - add set_primary method in Integration API ([cfbb7cd](https://github.com/novuhq/novu-python/commit/cfbb7cd3388bb6d604b4d645c9354e75b5698953))
+
+# [1.7.0](https://github.com/novuhq/novu-python/compare/v1.6.0...v1.7.0) (2023-10-14)
+
+
+### Features
+
+* **blueprint:** add blueprint api ([0e885ea](https://github.com/novuhq/novu-python/commit/0e885ea55c9fec7cd70892f4c1935e92f15e1e21))
+* **subscriber:** [#59](https://github.com/novuhq/novu-python/issues/59) - add bulk create for subscribers ([9b5983d](https://github.com/novuhq/novu-python/commit/9b5983d506a020b0361bc03d218196602aa3c857))
+* **subscriber:** added the delete_credentials method to the SubscriberApi ([d6ffc98](https://github.com/novuhq/novu-python/commit/d6ffc98c0c448cf6bbe8326f3c3864a8431d0c50))
+
+# [1.6.0](https://github.com/novuhq/novu-python/compare/v1.5.0...v1.6.0) (2023-10-14)
+
+
+### Features
+
+* **event:** [#94](https://github.com/novuhq/novu-python/issues/94) - support tenant in event api ([50116d6](https://github.com/novuhq/novu-python/commit/50116d68a7475f56ca5f74e37e5de397fd86abd3))
+
+# [1.5.0](https://github.com/novuhq/novu-python/compare/v1.4.0...v1.5.0) (2023-10-14)
+
+
+### Bug Fixes
+
+* **change:** promoted path parms is required for listing ([#86](https://github.com/novuhq/novu-python/issues/86)) ([9758660](https://github.com/novuhq/novu-python/commit/975866055a0aaab201d27c7fb7f466dd57e46a35))
+* **notification:** list notifications should return a paginated response ([#87](https://github.com/novuhq/novu-python/issues/87)) ([1b47e10](https://github.com/novuhq/novu-python/commit/1b47e10deee9be3a015e826d8111e5e8f312e0fb)), closes [#88](https://github.com/novuhq/novu-python/issues/88)
+
+
+### Features
+
+* added missing methods in notification group class ([#83](https://github.com/novuhq/novu-python/issues/83)) ([951d893](https://github.com/novuhq/novu-python/commit/951d8938ceaca68e211ed3991c0efb93f543b72b))
+* **subscriber:** add channels fields on DTO ([#53](https://github.com/novuhq/novu-python/issues/53)) ([b8809b5](https://github.com/novuhq/novu-python/commit/b8809b5b4b342f4fff15fa515b60599d3079fdba))
+
 # [1.4.0](https://github.com/novuhq/novu-python/compare/v1.3.1...v1.4.0) (2023-09-13)
 
 
@@ -9,7 +107,7 @@
 
 ### Features
 
-* **#9:** add notifications ressource in wrappers ([f33a32f](https://github.com/novuhq/novu-python/commit/f33a32fc788085516515ba7a3fa54181248e43bf)), closes [#9](https://github.com/novuhq/novu-python/issues/9)
+* **#9:** add notifications resource in wrappers ([f33a32f](https://github.com/novuhq/novu-python/commit/f33a32fc788085516515ba7a3fa54181248e43bf)), closes [#9](https://github.com/novuhq/novu-python/issues/9)
 * **tenant:** introduce Tenant API ([330cb19](https://github.com/novuhq/novu-python/commit/330cb19a679f1284487acd8f6beabd3edc681448))
 
 ## [1.3.1](https://github.com/novuhq/novu-python/compare/v1.3.0...v1.3.1) (2023-08-30)